Require explicit codec for endpoints
This replaces the previous solution which was serialising `Irmin_unix`
URI strings to be passed to be used in constructing remotes at the
server (`?headers:Cohttp.Header.t -> string -> remote`). The new
mechanism is more generic, allowing e.g. Conduit endpoints to be used as
remote endpoints rather than URI strings.

There are two limitations remaining with the current solution:

- `Git_mirage.endpoint` is not currently serialisable, as it contains a
  `Conduit_mirage.t`. As a result, there is no convenient way to use RPC
  to speak to a remote `Irmin_mirage_git` store (this was already the case).

- The CLI "resolvers" being used in `irmin-rpc-unix/bin/main.ml`
  (piggy-backed from `Irmin_unix`) are no longer sufficient to provide
  access to the `SYNC` API. A solution here would be to extend/modify
  the resolvers in `mirage/irmin`, but that's not a priority.