Refactor the server API
Previously, the server API was `Server.{create,run}`, where `create`
actually spawns the service asynchronously and `run` just waits forever
on an abstract type returned by `create`. This suggests that the service
only begins at the point when `run` is actually called (which is not
easily achievable with the Capnp API).

The proposed API is `Server.serve`, which creates and 'runs' the server
in one call (matching the underlying Capnp API for serving vats). I've
added some documentation explaining the semantics while I'm at it.